moved fetching to ToDoList
